Medicy/requirements.txt
pillow
tk
mysql-connector-python